feat(frontend): edit satellite details
